feat: (ImageViewer) support `ImageViewer.clear()`
