+++ src/components/image-viewer/index.ts
@@ -1,11 +1,11 @@
 import './image-viewer.less'
 import { attachPropertiesToComponent } from '../../utils/attach-properties-to-component'
+import { ImageViewer, MultiImageViewer } from './image-viewer'
 import {
-  ImageViewer,
-  MultiImageViewer,
-  showImageViewer,
   showMultiImageViewer,
-} from './image-viewer'
+  showImageViewer,
+  clearImageViewer,
+} from './methods'
 
 export type { ImageViewerProps, MultiImageViewerProps } from './image-viewer'
 
@@ -16,4 +16,5 @@ const Multi = attachPropertiesToComponent(MultiImageViewer, {
 export default attachPropertiesToComponent(ImageViewer, {
   Multi,
   show: showImageViewer,
+  clear: clearImageViewer,
 })
